Remove redundat decoration, optimize HeaderItemDecoration
